requirements.txt
streamlit==1.31.1
pandas==2.1.4
pydantic==2.5.3
langchain==0.1.6
langchain-core==0.1.23
langgraph==0.0.26
databricks-langchain==0.1.0
python-dotenv==1.0.0
typing-extensions==4.9.0